Split permissions module into separate modules for both platforms
